requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.14.42
botocore==1.17.42
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-crispy-forms==1.9.2
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
kombu==4.6.11
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
